work/drawing-checker/requirements.txt
PyMuPDF>=1.24
ezdxf>=1.3
pytesseract>=0.3.10
opencv-python>=4.10
pillow>=10.4
pywin32>=306; platform_system=="Windows"
google-genai>=1.0
python-dotenv>=1.0
jinja2>=3.1
tkinterdnd2>=0.4
